backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-seed==0.3.1
djangorestframework==3.12.4
Faker==13.11.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toposort==1.7
typing-extensions==3.10.0.2
